+++ 01-node-tutorial/answers/02-globals.js
