+++ 01-node-tutorial/answers/03-modules.js
@@ -1,0 +1,9 @@
+const names = require('./04-names.js');
+const sayHello = require('./05-utils.js');
+const data = require('./06-alternative-flavor.js');
+require('./07-mind-grenade.js');
+sayHello('susan')
+sayHello(names.anna)
+sayHello(names.maria)
+
+console.log(data);
