+++ 01-node-tutorial/answers/06-alternative-flavor.js
