+++ 01-node-tutorial/answers/07-mind-grenade.js
@@ -1,0 +1,8 @@
+const num1 = 10
+const num2 = 20
+
+function addValues() {
+  console.log(`The sum is : ${num1 + num2}`)
+}
+
+addValues()
