--- 01-node-tutorial/answers/mind-grenade.js
@@ -1,8 +1,0 @@
-const num1 = 5
-const num2 = 10
-
-function addValues() {
-  console.log(`the sum is : ${num1 + num2}`)
-}
-
-addValues()
