--- 01-node-tutorial/answers/modules.js
@@ -1,9 +1,0 @@
-const names = require('./names.js');
-const sayHello = require('./utils');
-const data = require('./alternative-flavor');
-require('./mind-grenade');
-sayHello('susan')
-sayHello(names.anna)
-sayHello(names.maria)
-
-console.log(data);
